pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-futures-backtester"
version = "0.1.0"
description = "High-performance vectorized futures backtesting framework built on VectorBT"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Simple Futures Backtester Team"}
]
keywords = [
    "backtesting",
    "trading",
    "futures",
    "vectorbt",
    "quantitative-finance",
    "algorithmic-trading"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]

dependencies = [
    # CLI and console
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
    # Configuration
    "pyyaml>=6.0,<7.0",
    # Data handling
    "pandas>=2.0,<3.0",
    "numpy>=1.24,<2.0",
    # Backtesting framework
    "vectorbt>=0.26.0",
    # JIT compilation
    "numba>=0.58,<1.0",
    # Visualization
    "plotly>=5.18,<6.0",
    "kaleido>=0.2.1,<0.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4,<9.0",
    "pytest-cov>=4.1,<6.0",
    "pytest-benchmark>=4.0,<5.0",
    # Code quality
    "black>=23.0,<25.0",
    "ruff>=0.1,<1.0",
    # Type checking
    "mypy>=1.5,<2.0",
    "pandas-stubs>=2.0,<3.0",
]

[project.scripts]
sfb = "simple_futures_backtester.cli:main"

[project.urls]
Homepage = "https://github.com/simple-futures-backtester/simple-futures-backtester"
Documentation = "https://github.com/simple-futures-backtester/simple-futures-backtester#readme"
Repository = "https://github.com/simple-futures-backtester/simple-futures-backtester.git"
Issues = "https://github.com/simple-futures-backtester/simple-futures-backtester/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["simple_futures_backtester*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | lib/vectorbt
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "lib/vectorbt",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["simple_futures_backtester"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "lib/vectorbt",
    "tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "bench_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as benchmarks (deselect with '-m \"not benchmark\"')",
    "bars: marks tests as bar generation benchmarks",
    "backtest: marks tests as backtest/sweep benchmarks",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["simple_futures_backtester"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "lib/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
